renee-data debt payment
TIP: I would pull this branch down and not bother
with looking over the diff on github since moving
so much around will make the diff pretty
unintelligible.

QUESTION: Could data.go be moved to renee-data
-api or go-renee and be imported from there to
other projects? I (Noah) would like to define
methods on these types, but since they are
imported, Go won't let you define methods on them
in other projects. I'm thinking of methods for
group signups, for sending messages, etc.

METADATA: I created three types to contain the
most common metadata found among our types:
Metadata (CreatedAt, LastUpdated, UUID),
AccountMetadata (AccountID, AccountName), and
CampusMetadata (CampusID, CampusName).

TODO: rename Account -> AccountID and Campus ->
CampusID on every object in the DB. This rename
will make our code more readable as it will tell
the writer exactly what field they are reading
from the object. Before we would write
campus.Account which could be misunderstood to
mean we wanted to read an Account object off of
the campus object.

ACCOUNT: The Account type is composed of Metadata
and the Name field.

CAMPUS: The Campus type is composed of
AccountMetadata, Metadata, and its remaining
fields. Since there were so many fields, I
organized them according to their relationships.

MEMBER: The Member type is composed of
AccountMetadata, CampusMetadata, Metadata, and
its remaining fields. I also organized the
remaining fields according to their relationships
I also adopted the standard that any field that
not a basic type should have the suffix "Data"
for the same reason as the name change mentioned
above under METADATA. This will make our code
much more readable. There were a few of these
"Data" fields that were all different types but
had the same fields, so I consolidated them all
into the new type StatusData (CreatedAt and
Status).

TODO: Append "Data" onto the names of the
appropriate fields in the DB and rename TIH
and DYP Data fields to match GrowthTrackW1Data
and GrowthTrackW2Data (As much as we can I'd
like to move away from using any PC related names
in our code)

MEMBER COMPOSITIONAL TYPES: these are the types
that make up the Member type. Here I mostly
documented more thoroughly and followed the same
"Data" naming standard I set above. This is where
I defined the StatusData type to consolidate
other types having the same fields.

GROUP: The Group type is composed of
AccountMetadata, CampusMetadata, Metadata, and
its remaining fields. I organized the fields.
Here, I strayed from the "Data" naming convention
with the exception of renaming Capacity ->
CapacityData. This is because with names like
Details, Leaders, Oversights, Registrations, and
Attendance, it is acceptably clear to me what is
to be expected from reading those fields.
(Though, the Attendance field could be renamed
AttendanceDateToAttendanceList or
MapDateToAttendance, but I'll leave that up to
the next person to seek to pay down debt in this
file.) I then went on to define the types
Event, GrowGroup, GrowthTrack, InterestGroup, and
ServeTeam. These types are what I see as the most
valuable gain from this exercise. If the answer
to my QUESTION at the top of this commit is "Yes"
then these types will allow us to move away from
the gross use of the "Group" type all over the
place in our code and to replace it with an
appropriately named type from that list and to
define methods on these types. These new types
also contain fields from the original Group type
that were only used for that specific type of
Group, which I think is another extremely
valuable add for the readability of our code. We
should not be able to read the tier field from an
Event or the StudyKey field from an InterstGroup.
(Although, in the future when we want to build
attendance tracking for ALL groups, then we can
make a general AttendanceKey field on Group and
remove StudyKey from GrowGroup)

TODO: Rename Capacity to CapacityData in the DB
for all Groups.

GROUP CONSTS: The only valuable group-related
consts I could think to define were the different
types of groups. So, I defined GroupType and all
of the currently used GroupTypes. I also defined
IsValid on GroupType so that users can easily
check if their provided type is valid. I see this
as the second most valuable gain from this
commit. If you're interested in seeing this in
action go here:
https://play.golang.org/p/JaVlE9Awknh

TODO: Use the GroupTypes when constructing any
Group and assigning its type and use IsValid in
the API to check incoming group types.

GROUP COMPOSITIONAL TYPES: these are the types
that make up the Group type. More documentation.

TODO: move tier from Details to root object for
grow groups in DB.

MISC/UNUSED: NewCreationQR is put here since it
is not used by another type in the file.